// src/app/athlete/athlete.store.dto.ts
export interface AthleteStoreDto {
    id: string;
    username: string;
    firstname: string;
    lastname: string;
    city: string;
    state: string;
    country: string;
    sex: string;
    premium: boolean;
    summit: boolean;
    createdAt: string;
    updatedAt: string;
    profileMedium: string;
    profile: string;
}